Application of creating logging file in a program
This is a program allows you to print special characters to create a triangle shape. However, one special thing is I incorporate the use of diary function to create logging file. Logging file is another way to assist you during debugging process. I just thought it useful so include this in this repository.